Rename *.hpp files to *.h
