+++ engine/src/audio/renderers/OpenAL/OpenALRenderableStreamingSource.cpp
@@ -27,7 +27,7 @@
 //
 // C++ Implementation: Audio::OpenALRenderableListener
 //
-#include "vega_cast_utils.hpp"
+#include "vega_cast_utils.h"
 #include "OpenALRenderableStreamingSource.h"
 #include "OpenALStreamingSound.h"
 #include "OpenALHelpers.h"
