+++ engine/src/cmd/collide2/Stdafx.h
@@ -46,7 +46,7 @@
 #include "opcodetypes.h"
 #include "opcodesysdef.h"
 #include "Opcode.h"
-#include "vs_collections.hpp"
+#include "vs_collections.h"
 
 #endif    // _STDAFX_H_
