+++ engine/src/vega_py_run.h
@@ -1,7 +1,5 @@
 /*
- * vega_string_utils.hpp
- *
- * Copyright (C) 2001-2022 Daniel Horn, Stephen G. Tuggy,
+ * Copyright (C) 2001-2022 Daniel Horn, pyramid3d, Stephen G. Tuggy,
  * and other Vega Strike contributors.
  *
  * https://github.com/vegastrike/Vega-Strike-Engine-Source
@@ -22,29 +20,13 @@
  * along with Vega Strike. If not, see <https://www.gnu.org/licenses/>.
  */
 
-#ifndef VEGA_STRIKE_SRC_VEGA_STRING_UTILS_HPP_
-#define VEGA_STRIKE_SRC_VEGA_STRING_UTILS_HPP_
 
-#include "vs_logging.h"
+#ifndef VEGA_STRIKE_SRC_VEGA_PY_RUN_H_
+#define VEGA_STRIKE_SRC_VEGA_PY_RUN_H_
 
-inline char *vega_str_dup(const char *string) {
-#if _XOPEN_SOURCE >= 500 || _POSIX_C_SOURCE >= 200809L
-    return strdup(string);
-#elif defined (WIN32)
-    return _strdup(string);
-#else
-    size_t buf_size = strlen(string) + 1;
-    char *alloc;
-    alloc = (char *)malloc(buf_size);
-    if (alloc == nullptr)
-    {
-        VS_LOG_AND_FLUSH(fatal, "Out of memory");
-        VSExit(-1);
-    }
-    strncpy(alloc, string, buf_size);
-    alloc[buf_size - 1] = '\0';
-    return alloc;
-#endif
-}
+#include <string>
 
-#endif //VEGA_STRIKE_SRC_VEGA_STRING_UTILS_HPP_
+void VegaPyRunString(char * py_snippet);
+void VegaPyRunString(const std::string& py_snippet);
+
+#endif //VEGA_STRIKE_SRC_VEGA_PY_RUN_H_
